test: fix tests for browser
